% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aws_igf_GIANT149}
\alias{aws_igf_GIANT149}
\title{an indexedGF instance for 149 files in GIANT consortium portal, resident in AWS S3}
\format{indexedGF instance}
\usage{
aws_igf_GIANT149
}
\description{
an indexedGF instance for 149 files in GIANT consortium portal, resident in AWS S3
}
\note{
A small 'enrichment' dataset was dropped from the full
set of 150 files.  Metadata components include 'field_names'
and 'nsnps' which are coordinated with the 149 files but
have no element names.  Note that there has been light curation
relative to material at the GIANT portal, where some of the files
have Pax headers indicative of compression in tar archive
context, and some files are not compressed despite their suffix.
}
\keyword{datasets}
