% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igf_GIANT149}
\alias{igf_GIANT149}
\title{an indexedGF instance for 149 files in GIANT consortium portal}
\format{indexedGF instance}
\usage{
igf_GIANT149
}
\description{
an indexedGF instance for 149 files in GIANT consortium portal
}
\note{
A small 'enrichment' dataset was dropped from the full
set of 150 files.  Metadata components include 'field_names'
and 'nsnps' which are coordinated with the 149 files but
have no element names.  The filenames used here assume the
data are locally available in the current folder.  See aws_igf_GIANT149
for access to globally accessible images in AWS S3.
}
\keyword{datasets}
