% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igf_giant150}
\alias{igf_giant150}
\title{an indexedGF instance for 150 files in GIANT consortium portal}
\format{indexedGF instance}
\usage{
igf_giant150
}
\description{
an indexedGF instance for 150 files in GIANT consortium portal
}
\keyword{datasets}
