% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{indexedGF}
\alias{indexedGF}
\title{constructor for indexedGF}
\usage{
indexedGF(gfi, ...)
}
\arguments{
\item{gfi}{instance of GenomicFiles}
}
\description{
constructor for indexedGF
}
\examples{
indexedGF(indexedGF::WHRadjBMI_gf)
}
