% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{vroomIn}
\alias{vroomIn}
\title{populate an indexedGF by running vroom over the files() component}
\usage{
vroomIn(igf, itmeth = lapply)
}
\arguments{
\item{igf}{an indexedGF instance}

\item{itmeth}{a function, defaults to lapply}
}
\value{
An updated indexedGF instance; there may be try-error instances in the vrlist slot.
}
\description{
populate an indexedGF by running vroom over the files() component
}
\note{
For applications over HTTP parallelized ingestion can generate certification errors.
}
\examples{
nigf = indexedGF::WHRadjBMI_gf
nigf
tst = vroomIn(nigf)
tst
}
