% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WHRadjBMI_gf}
\alias{WHRadjBMI_gf}
\title{a GenomicFiles instance representing a piece of the GIANT consortium portal at Broad}
\format{GenomicFiles instance}
\source{
\url{https://portals.broadinstitute.org/collaboration/giant/index.php/GIANT_consortium_data_files}
}
\usage{
WHRadjBMI_gf
}
\description{
a GenomicFiles instance representing a piece of the GIANT consortium portal at Broad
}
\keyword{datasets}
