% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{whrurls}
\alias{whrurls}
\title{a set of URLS for broad institute portal on GIANT consortium}
\usage{
whrurls()
}
\description{
a set of URLS for broad institute portal on GIANT consortium
}
\examples{
head(whrurls(),2)
ww = whurls()
ss = strsplit(ww, "\\\\.")
table(sapply(ss, "[", 6))
}
